+++ packages/components/ui/src/sensors/screen/WindowBinder.svelte
@@ -1,0 +1,11 @@
+<script>
+	export let innerHeight;
+	export let innerWidth;
+	export let onResize;
+</script>
+
+<svelte:window
+	bind:innerWidth={innerWidth}
+	bind:innerHeight={innerHeight}
+	on:resize={onResize}
+/>