+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - add `ResponsiveFlex` component
 - `ScreenGauge`: fixed `sampleText` enlarging the document size
 - rename `ScreenGauge` to `ScreenSensor`
+- `ScreenSensor`: made it into a singleton so that it renders only once
 
 ## `@svizzle/atlas` v0.6.0 (next)