+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 	- fixed `sampleText` enlarging the document size
 	- renamed `ScreenGauge` to `ScreenSensor`
 	- made `ScreenSensor` into a singleton so that it renders only once
+- add `StorageIO` component
 
 ## `@svizzle/atlas` v0.6.0 (next)